Pass stdin-path to phpcbf
Since 232cfba the validator passes the filename to phpcs, so it respects all include and exclude patterns. However, when the code is formatted the filename isn't passed, which led to unexpected results.

This makes the same change that was made to phpcs to phpcbf, so they both behave the same and autofixing files takes into account the include/exclude rules.